requirements.txt
# Core & Data Processing
pandas>=1.5.0
numpy==1.23.5
requests>=2.28.0
python-dateutil>=2.8.0
loguru>=0.7.0

# Exchange Connector
ccxt>=4.0.0

# Technical Analysis & Features
pandas-ta==0.3.14b0

# Machine Learning
xgboost>=1.6.0
scikit-learn>=1.1.0
joblib>=1.3.0

# Hyperparameter Optimization
optuna>=3.0.0

# Backtesting & Plotting
Backtesting>=0.3.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Telegram Bot
python-telegram-bot==20.8

# Configuration
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=0.19.0

# Database
SQLAlchemy>=2.0.0

# Asyncio (required by some libraries)
nest_asyncio>=1.6.0


setuptools>=69.0.0